chore: move tutorial demo files out of src/poc into src/shared/tutorial
These files (AnimatedDemo, constants, utils) back the live "How to Play"
modal and aren't prototype code, despite living under poc/.
